チェックリスト登録機能実装

+++ db/migrate/20220221135102_create_problems.rb
@@ -3,12 +3,12 @@ class CreateProblems < ActiveRecord::Migration[6.0]
     create_table :problems do |t|
       t.string     :name,            null: false
       t.integer    :grade_id,        null: false
-      t.float      :holding,         null: false
-      t.float      :physical,        null: false
-      t.float      :move,            null: false
-      t.float      :positioning,     null: false
-      t.float      :footwork,        null: false
-      t.float      :coordination,    null: false
+      t.integer    :holding,         null: false
+      t.integer    :physical,        null: false
+      t.integer    :move,            null: false
+      t.integer    :positioning,     null: false
+      t.integer    :footwork,        null: false
+      t.integer    :coordination,    null: false
       t.references :admin,           null: false, foreign_key: true
       t.references :problem_list,    null: false, foreign_key: true
     end
